Added a deactivation controller method to UserProfileController
